refactor: remove dead diagnose flow and split site store
Drop the unused preflight diagnostics path and legacy helpers so the active CLI surface stays smaller. Break site store code and tests into focused files to make future changes safer.
