+++ cmd/deck/command_handlers.go
@@ -22,7 +22,6 @@ import (
 
 	"github.com/taedi90/deck/internal/bundle"
 	"github.com/taedi90/deck/internal/config"
-	"github.com/taedi90/deck/internal/diagnose"
 	"github.com/taedi90/deck/internal/install"
 	ctrllogs "github.com/taedi90/deck/internal/logs"
 	"github.com/taedi90/deck/internal/nodeid"
@@ -2276,63 +2275,6 @@ func readLogsFile(path string) ([]ctrllogs.LogRecord, error) {
 	return records, nil
 }
 
-func resolveControlLogFilePath(cliPath string) (string, error) {
-	if cliPath != "" {
-		if _, err := os.Stat(cliPath); err != nil {
-			if os.IsNotExist(err) {
-				return "", fmt.Errorf("control logs: log file not found: %s", cliPath)
-			}
-			return "", fmt.Errorf("control logs: stat log file: %w", err)
-		}
-		return cliPath, nil
-	}
-	candidates := []string{
-		filepath.Join(".", "bundle", ".deck", "logs", "server-audit.log"),
-		filepath.Join(".", ".deck", "logs", "server-audit.log"),
-	}
-	for _, candidate := range candidates {
-		if _, err := os.Stat(candidate); err == nil {
-			return candidate, nil
-		}
-	}
-	return "", errors.New("control logs: server audit log not found (use --path)")
-}
-
-func readControlLogsFile(path string, since time.Duration) ([]ctrllogs.LogRecord, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return nil, fmt.Errorf("control logs: open log file: %w", err)
-	}
-	defer f.Close()
-
-	var cutoff time.Time
-	if since > 0 {
-		cutoff = time.Now().Add(-since)
-	}
-
-	records := []ctrllogs.LogRecord{}
-	scanner := bufio.NewScanner(f)
-	scanner.Buffer(make([]byte, 0, 64*1024), 1024*1024)
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if line == "" {
-			continue
-		}
-		record, parseErr := ctrllogs.NormalizeJSONLine([]byte(line))
-		if parseErr != nil {
-			continue
-		}
-		if since > 0 && !recordAfterCutoff(record, cutoff) {
-			continue
-		}
-		records = append(records, record)
-	}
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("control logs: read log file: %w", err)
-	}
-	return records, nil
-}
-
 func readControlLogsJournal(unit string, tail int, since time.Duration) ([]ctrllogs.LogRecord, error) {
 	args := []string{"-u", unit, "-o", "json", "--no-pager", "-n", strconv.Itoa(tail)}
 	if since > 0 {
@@ -2343,108 +2285,6 @@ func readControlLogsJournal(unit string, tail int, since time.Duration) ([]ctrll
 		return nil, classifyJournalctlError(err, strings.TrimSpace(string(raw)))
 	}
 	return parseJournalOutputLines(raw), nil
-}
-
-func followControlLogsJournal(unit string, tail int, since time.Duration, filters ctrllogs.LogFilters, output string) error {
-	args := []string{"-u", unit, "-o", "json", "--no-pager", "-n", strconv.Itoa(tail), "-f"}
-	if since > 0 {
-		args = append(args, "--since", formatJournalSince(since))
-	}
-	cmd := exec.Command("journalctl", args...)
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return fmt.Errorf("control logs: journal stdout: %w", err)
-	}
-	stderr, err := cmd.StderrPipe()
-	if err != nil {
-		return fmt.Errorf("control logs: journal stderr: %w", err)
-	}
-	if err := cmd.Start(); err != nil {
-		base := classifyJournalctlError(err, "")
-		return fmt.Errorf("control logs: %w\nsuggestion: %s", base, suggestJournalctlCommand(unit))
-	}
-
-	stderrCh := make(chan string, 1)
-	go func() {
-		buf, _ := io.ReadAll(stderr)
-		stderrCh <- strings.TrimSpace(string(buf))
-	}()
-
-	scanner := bufio.NewScanner(stdout)
-	scanner.Buffer(make([]byte, 0, 64*1024), 1024*1024)
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if line == "" {
-			continue
-		}
-		var raw map[string]any
-		if err := json.Unmarshal([]byte(line), &raw); err != nil {
-			continue
-		}
-		record := ctrllogs.NormalizeJournalRecord(raw)
-		if !ctrllogs.MatchesLogFilters(record, filters) {
-			continue
-		}
-		if err := printControlLogRecord(record, output); err != nil {
-			return err
-		}
-	}
-	if scanErr := scanner.Err(); scanErr != nil {
-		return fmt.Errorf("control logs: follow journal: %w", scanErr)
-	}
-	waitErr := cmd.Wait()
-	stderrText := <-stderrCh
-	if waitErr != nil {
-		base := classifyJournalctlError(waitErr, stderrText)
-		return fmt.Errorf("control logs: %w\nsuggestion: %s", base, suggestJournalctlCommand(unit))
-	}
-	return nil
-}
-
-func followControlLogsFile(path string, filters ctrllogs.LogFilters, output string) error {
-	info, err := os.Stat(path)
-	if err != nil {
-		return fmt.Errorf("control logs: stat log file: %w", err)
-	}
-	offset := info.Size()
-	ticker := time.NewTicker(1 * time.Second)
-	defer ticker.Stop()
-
-	for range ticker.C {
-		f, err := os.Open(path)
-		if err != nil {
-			continue
-		}
-		if _, err := f.Seek(offset, 0); err != nil {
-			_ = f.Close()
-			continue
-		}
-		scanner := bufio.NewScanner(f)
-		scanner.Buffer(make([]byte, 0, 64*1024), 1024*1024)
-		for scanner.Scan() {
-			line := strings.TrimSpace(scanner.Text())
-			if line == "" {
-				continue
-			}
-			record, parseErr := ctrllogs.NormalizeJSONLine([]byte(line))
-			if parseErr != nil {
-				continue
-			}
-			if !ctrllogs.MatchesLogFilters(record, filters) {
-				continue
-			}
-			if err := printControlLogRecord(record, output); err != nil {
-				_ = f.Close()
-				return err
-			}
-		}
-		nextOffset, err := f.Seek(0, 1)
-		if err == nil {
-			offset = nextOffset
-		}
-		_ = f.Close()
-	}
-	return nil
 }
 
 func parseJournalOutputLines(raw []byte) []ctrllogs.LogRecord {
@@ -2485,67 +2325,6 @@ func suggestJournalctlCommand(unit string) string {
 
 func formatJournalSince(since time.Duration) string {
 	return time.Now().Add(-since).Format(time.RFC3339)
-}
-
-func recordAfterCutoff(record ctrllogs.LogRecord, cutoff time.Time) bool {
-	ts := strings.TrimSpace(record.TS)
-	if ts == "" {
-		return true
-	}
-	parsed, err := time.Parse(time.RFC3339Nano, ts)
-	if err != nil {
-		return true
-	}
-	return !parsed.Before(cutoff)
-}
-
-func filterControlLogRecords(records []ctrllogs.LogRecord, filters ctrllogs.LogFilters) []ctrllogs.LogRecord {
-	filtered := make([]ctrllogs.LogRecord, 0, len(records))
-	for _, record := range records {
-		if ctrllogs.MatchesLogFilters(record, filters) {
-			filtered = append(filtered, record)
-		}
-	}
-	return filtered
-}
-
-func tailControlLogRecords(records []ctrllogs.LogRecord, tail int) []ctrllogs.LogRecord {
-	if len(records) <= tail {
-		return records
-	}
-	return records[len(records)-tail:]
-}
-
-func printControlLogRecords(records []ctrllogs.LogRecord, output string) error {
-	if output == "json" {
-		encoder := json.NewEncoder(os.Stdout)
-		for _, record := range records {
-			if err := encoder.Encode(record); err != nil {
-				return fmt.Errorf("control logs: encode output: %w", err)
-			}
-		}
-		return nil
-	}
-	for _, record := range records {
-		if _, err := fmt.Fprintln(os.Stdout, ctrllogs.FormatLogText(record)); err != nil {
-			return fmt.Errorf("control logs: write output: %w", err)
-		}
-	}
-	return nil
-}
-
-func printControlLogRecord(record ctrllogs.LogRecord, output string) error {
-	if output == "json" {
-		if err := json.NewEncoder(os.Stdout).Encode(record); err != nil {
-			return fmt.Errorf("control logs: encode output: %w", err)
-		}
-		return nil
-	}
-	_, err := fmt.Fprintln(os.Stdout, ctrllogs.FormatLogText(record))
-	if err != nil {
-		return fmt.Errorf("control logs: write output: %w", err)
-	}
-	return nil
 }
 
 func isPermissionError(msg string) bool {
@@ -3524,125 +3303,5 @@ func runValidate(args []string) error {
 	}
 
 	fmt.Fprintln(os.Stdout, "validate: ok")
-	return nil
-}
-
-func runRun(args []string) error {
-	fs := flag.NewFlagSet("run", flag.ContinueOnError)
-	fs.SetOutput(os.Stderr)
-
-	var file string
-	registerFileFlags(fs, &file, "path or URL to workflow file")
-	phase := fs.String("phase", "", "phase to execute (prepare/install)")
-	bundle := fs.String("root", "", "bundle output path")
-	vars := &varFlag{}
-	fs.Var(vars, "var", "set variable override (key=value), repeatable")
-	if err := fs.Parse(args); err != nil {
-		return err
-	}
-
-	if file == "" {
-		return errors.New("--file (or -f) is required")
-	}
-	if *phase == "" {
-		return errors.New("--phase is required")
-	}
-
-	if err := validate.File(file); err != nil {
-		return err
-	}
-
-	wf, err := config.LoadWithOptions(file, config.LoadOptions{VarOverrides: varsAsAnyMap(vars.AsMap())})
-	if err != nil {
-		return err
-	}
-
-	switch *phase {
-	case "prepare":
-		if err := prepare.Run(wf, prepare.RunOptions{BundleRoot: *bundle}); err != nil {
-			return err
-		}
-		fmt.Fprintln(os.Stdout, "run prepare: ok")
-		return nil
-	case "install":
-		if err := install.Run(wf, install.RunOptions{BundleRoot: *bundle}); err != nil {
-			return err
-		}
-		fmt.Fprintln(os.Stdout, "run install: ok")
-		return nil
-	default:
-		return fmt.Errorf("unsupported phase: %s", *phase)
-	}
-}
-
-func runResume(args []string) error {
-	fs := flag.NewFlagSet("resume", flag.ContinueOnError)
-	fs.SetOutput(os.Stderr)
-
-	var file string
-	registerFileFlags(fs, &file, "path or URL to workflow file")
-	bundle := fs.String("root", "", "bundle path")
-	vars := &varFlag{}
-	fs.Var(vars, "var", "set variable override (key=value), repeatable")
-	if err := fs.Parse(args); err != nil {
-		return err
-	}
-	if file == "" {
-		return errors.New("--file (or -f) is required")
-	}
-
-	if err := validate.File(file); err != nil {
-		return err
-	}
-
-	wf, err := config.LoadWithOptions(file, config.LoadOptions{VarOverrides: varsAsAnyMap(vars.AsMap())})
-	if err != nil {
-		return err
-	}
-	if err := install.Run(wf, install.RunOptions{BundleRoot: *bundle}); err != nil {
-		return err
-	}
-	fmt.Fprintln(os.Stdout, "resume install: ok")
-	return nil
-}
-
-func runDiagnose(args []string) error {
-	fs := flag.NewFlagSet("diagnose", flag.ContinueOnError)
-	fs.SetOutput(os.Stderr)
-
-	var file string
-	registerFileFlags(fs, &file, "path or URL to workflow file")
-	bundle := fs.String("root", "", "bundle path")
-	preflight := fs.Bool("preflight", false, "run preflight checks")
-	out := fs.String("out", "reports/diagnose.json", "diagnose report output path")
-	vars := &varFlag{}
-	fs.Var(vars, "var", "set variable override (key=value), repeatable")
-	if err := fs.Parse(args); err != nil {
-		return err
-	}
-	if file == "" {
-		return errors.New("--file (or -f) is required")
-	}
-	if !*preflight {
-		return errors.New("only --preflight mode is supported")
-	}
-
-	if err := validate.File(file); err != nil {
-		return err
-	}
-
-	wf, err := config.LoadWithOptions(file, config.LoadOptions{VarOverrides: varsAsAnyMap(vars.AsMap())})
-	if err != nil {
-		return err
-	}
-
-	report, err := diagnose.Preflight(wf, diagnose.RunOptions{WorkflowPath: file, BundleRoot: *bundle, OutputPath: *out, EnforceHostChecks: true})
-	if err != nil {
-		fmt.Fprintf(os.Stdout, "diagnose preflight: failed (%d failed checks)\n", report.Summary.Failed)
-		return err
-	}
-
-	fmt.Fprintf(os.Stdout, "diagnose preflight: ok (%d checks)\n", report.Summary.Passed)
-	fmt.Fprintf(os.Stdout, "diagnose report: %s\n", *out)
 	return nil
 }
